+++ src/main/java/expecti/Evaluator.java
@@ -8,26 +8,26 @@ public class Evaluator implements Evaluation
      * construct the Object used for getting the evaluation values
      */
     private final FitnessFunction FUNCTION = new FitnessFunction();
-    private int[] weights;
+    private double[] weights;
 
     /**
      * final evaluation, including the weights
      */
-    public int evaluation;
+    public double evaluation;
 
 
     /**
      * The constructor of the Evaluator object, this object is used solely to get the final evaluation value
      * @param game the Dicechess game at the current node
      */
-    Evaluator(DiceChess game, int[] weights)
+    Evaluator(DiceChess game, double[] weights)
     {
     evaluation = FUNCTION.evaluate(game,weights);
     }
 
 
     @Override
-    public int getEvaluation()
+    public double getEvaluation()
     {
         return evaluation;
     }
